chore(github): delete dead in-tree poller machinery (GH-4169)
Delete internal/adapters/github/project_source.go (in-tree board-source
read path, superseded by studio-sdk's ProjectBoardSource in #4168) and
spec_validator.go, along with the legacy in-tree issue handler path
(handleGitHubIssueWithResult and its exclusive helpers, plus the now-orphaned
applySpecGuard/spec_guard.go) that was spec_validator.go's sole caller.
internal/ghissue.ValidateSpec is now the single spec-guard implementation.

poller.go, cleanup.go, merger.go, retry.go, issue_creator.go, and
issue_create.go were left in place: each still has a live caller in
cmd/pilot/main.go (stale-label cleanup, sub-issue merge-wait, bot issue
creation) or internal/adapters/github/client.go (retry.go's HTTP retry
wrapper) that no sibling GH-4155 sub-issue removes — deleting them would
break the build and drop working features outside this subtask's scope
fence. Rationale documented in
.agent/sops/integrations/github-poller-dead-code-cascade-gh4169.md.
